Big db overhaul and other misc
