+++ alembic.ini
@@ -1,14 +1,13 @@
-# A generic, single database configuration.
-
 [alembic]
 # path to migration scripts
-script_location = .\app\core\db\migrations
+script_location = alembic
 
 # template used to generate migration files
 # file_template = %%(rev)s_%%(slug)s
 
 # sys.path path, will be prepended to sys.path if present.
 # defaults to the current working directory.
+# (new in 1.5.5)
 prepend_sys_path = .
 
 # timezone to use when rendering the date within the migration file
@@ -33,34 +32,34 @@ prepend_sys_path = .
 # sourceless = false
 
 # version location specification; This defaults
-# to .\app\core\db\migrations/versions.  When using multiple version
+# to ${script_location}/versions.  When using multiple version
 # directories, initial revisions must be specified with --version-path.
-# The path separator used here should be the separator specified by "version_path_separator"
-# version_locations = %(here)s/bar:%(here)s/bat:.\app\core\db\migrations/versions
+# The path separator used here should be the separator specified by "version_path_separator" below.
+# version_locations = %(here)s/bar:%(here)s/bat:${script_location}/versions
 
 # version path separator; As mentioned above, this is the character used to split
-# version_locations. Valid values are:
+# version_locations. The default within new alembic.ini files is "os", which uses os.pathsep.
+# If this key is omitted entirely, it falls back to the legacy behavior of splitting on spaces and/or commas.
+# Valid values for version_path_separator are:
 #
 # version_path_separator = :
 # version_path_separator = ;
 # version_path_separator = space
-version_path_separator = os  # default: use os.pathsep
+version_path_separator = os  # Use os.pathsep. Default configuration used for new projects.
 
 # the output encoding used when revision files
 # are written from script.py.mako
 # output_encoding = utf-8
 
-# THIS SHOULD BE CHANGED
 # sqlalchemy.url = driver://user:pass@localhost/dbname
-sqlalchemy.url = sqlite:///./app.db
 
-
-[post_write_hooks]
-# post_write_hooks defines scripts or Python functions that are run
+# [post_write_hooks]
+# This section defines scripts or Python functions that are run
 # on newly generated revision scripts.  See the documentation for further
 # detail and examples
 
-# format using "black" - use the console_scripts runner, against the "black" entrypoint
+# format using "black" - use the console_scripts runner,
+# against the "black" entrypoint
 # hooks = black
 # black.type = console_scripts
 # black.entrypoint = black
