--- app/core/db/database.py
@@ -1,12 +1,0 @@
-from sqlalchemy.ext.asyncio import create_async_engine, AsyncSession
-from sqlalchemy.ext.declarative import declarative_base
-from sqlalchemy.orm import sessionmaker
-
-SQLALCHEMY_DATABASE_URL = "sqlite:///./app.db"
-
-engine = create_async_engine(
-    SQLALCHEMY_DATABASE_URL, connect_args={"check_same_thread": False}
-)
-session_local = sessionmaker(autocommit=False, class_=AsyncSession, autoflush=False, bind=engine)
-
-Base = declarative_base()
