--- app/core/dependencies.py
@@ -1,1 +1,0 @@
-from fastapi import Header, HTTPException
