+++ app/gpt/models.py
@@ -1,0 +1,1 @@
+gpt_models = []
