--- app/sukima/gptauto.py
@@ -1,6 +1,0 @@
-class GPTAuto:
-        def __init__(self, model_name='generic'):
-                self.model_name = model_name
-        
-        def generate(self, prompt, args):
-                return prompt
