--- app/sukima/models.py
@@ -1,1 +1,0 @@
-gpt_models = []
