--- config.json
@@ -1,7 +1,0 @@
-{
-    "host": "0.0.0.0",
-    "port": 8080,
-    "auth_enable": true,
-    "auth_admin_token": "admin",
-    "auth_db": "db.json"
-}
