+++ docker-compose.yaml
@@ -1,0 +1,28 @@
+version: "3.8"
+
+services:
+    app:
+        build:
+            context: ./
+            dockerfile: Dockerfile
+        env_file:
+            - conf.env
+        ports:
+            - "8000:8000"
+        volumes:
+            - ./app/:/app/
+        command: uvicorn app.main:app --reload --host 0.0.0.0 --port 8000
+        depends_on:
+            - database
+
+    database:
+        image: postgres:14-bullseye
+        env_file:
+            - conf.env
+        ports:
+            - 5432:5432
+        volumes:
+            - postgres_data:/var/lib/postgresql/data/
+
+volumes:
+    postgres_data:
