+++ requirements.txt
@@ -1,11 +1,13 @@
 fastapi[all]
-alembic
 uvicorn[standard]
+databases[postgresql]
 python-jose[cryptography]
-python-multipart
+psycopg2
+asyncpg
+pydantic
+alembic
 sqlalchemy
-passlib
-bcrypt
-pyopenssl
-tinydb
+torch
 transformers
+bcrypt
+passlib
